backend/requirements.txt
fastapi==0.121.2
uvicorn==0.38.0
python-jose==3.5.0
cryptography==46.0.3
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
pydantic==2.12.4
python-dotenv==1.2.1
email-validator==2.3.0